pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestrator"
version = "0.1.0"
description = "AI Orchestrator — single-model SDLC automation"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "respx>=0.20",
    "httpx>=0.25.0",
    "pytest-asyncio>=0.23.0",
    "coverage>=7.0",
]

[project.scripts]
guildr = "orchestrator.cli:main"
orchestrate = "orchestrator.cli:main"

[tool.setuptools.packages.find]
include = ["orchestrator*", "web*"]

[tool.pytest.ini_options]
testpaths = ["tests", "web/backend/tests"]
